Add a modified version of TribitsPkgExportCacheVars.cmake (trilinos/Trilinos#11545)
This was copied from the TriBITS commit:

  commit 07aeef0f9f10d928b614214c1c8b38af0eecc6d4 (rab-github/551-fix-assert-enable-var-export)
  Author: Roscoe A. Bartlett <rabartl@sandia.gov>
  Date:   Mon Dec 12 19:20:48 2022 -0700

      Remove assert of local matches cache export var (#551)

and minor modifications were made to make it stand alone and remove the assert
function that is not needed.

This will be used to make the Kokkos native CMake build system produce the
correct set(Kokkos_<info> ???) statements in the installed KokkosConfig.cmake
file.