Pass in and define compilers before calling find_package(Trilinos) (#11545)
This makes the CUDA build work since KokkosConfig.cmake checks that the nvcc
compiler is being used when it is included.

This does not mean that customers still can't the compilers from
find_package(Trilinos), it is just that it will not work with CUDA.  This also
means we are no longer testing getting the compilers from Trilinos.  But given
that customers that use CUDA can't do that anyway, we should consider removing
the compilers from the TriBITS-generated <Package>Config.cmake and
<Project>Config.cmake files.